feat: ✨ useWagmiConnect composable
